Rename translations-extractor to translations
